Disable U8G2_16BIT with prebuild scrtipt to save memory + Clamp app.selected_param and better checks for its ranges + Hide chanel's pattern when steps == 1

+++ src/libGravity.cpp
@@ -69,7 +69,7 @@ void Gravity::Process() {
     cv2.Process();
 
     // Update Output states.
-    for (int i; i < OUTPUT_COUNT; i++) {
+    for (uint8_t i; i < OUTPUT_COUNT; i++) {
         outputs[i].Process();
     }
 }
